All the necessary files were uploaded.
# Edureka Projects

This repository contains projects that I have worked on during my time at Edureka.

## Project 1: Netflix Home Page Clone

The first project is a clone of the Netflix home page. It was designed using **HTML, CSS, and Bootstrap**. The focus of this project was on creating a visually appealing design. Please note that this project does not include any functionality.